+++ deeplearning4j-core/src/test/java/org/deeplearning4j/eval/EvaluationCalibrationTest.java
@@ -98,7 +98,8 @@ public class EvaluationCalibrationTest extends BaseDL4JTest {
         ec.eval(labels, arr);
 
         int[] expLabelCounts = labels.sum(0).data().asInt();
-        int[] expPredictionCount = new int[labels.size(1)];
+        // FIXME: int cast
+        int[] expPredictionCount = new int[(int) labels.size(1)];
         INDArray argmax = Nd4j.argMax(arr, 1);
         for (int i = 0; i < argmax.length(); i++) {
             expPredictionCount[argmax.getInt(i, 0)]++;
