+++ deeplearning4j-core/src/test/java/org/deeplearning4j/nn/layers/feedforward/dense/DenseTest.java
@@ -37,7 +37,7 @@ public class DenseTest extends BaseDL4JTest {
 
         NeuralNetConfiguration conf = new NeuralNetConfiguration.Builder().layer(build).build();
 
-        int numParams = conf.getLayer().initializer().numParams(conf);
+        long numParams = conf.getLayer().initializer().numParams(conf);
         INDArray params = Nd4j.create(1, numParams);
         Layer layer = conf.getLayer().instantiate(conf, null, 0, params, true);
