+++ deeplearning4j-core/src/test/java/org/deeplearning4j/nn/layers/samediff/TestSameDiffDense.java
@@ -49,8 +49,8 @@ public class TestSameDiffDense {
         assertNotNull(pt1.get(DefaultParamInitializer.WEIGHT_KEY));
         assertNotNull(pt1.get(DefaultParamInitializer.BIAS_KEY));
 
-        assertArrayEquals(new int[]{nIn, nOut}, pt1.get(DefaultParamInitializer.WEIGHT_KEY).shape());
-        assertArrayEquals(new int[]{1, nOut}, pt1.get(DefaultParamInitializer.BIAS_KEY).shape());
+        assertArrayEquals(new long[]{nIn, nOut}, pt1.get(DefaultParamInitializer.WEIGHT_KEY).shape());
+        assertArrayEquals(new long[]{1, nOut}, pt1.get(DefaultParamInitializer.BIAS_KEY).shape());
     }
 
     @Test
