+++ deeplearning4j-data/deeplearning4j-utility-iterators/src/main/java/org/deeplearning4j/datasets/iterator/file/FileMultiDataSetIterator.java
@@ -136,7 +136,8 @@ public class FileMultiDataSetIterator extends BaseFileIterator<MultiDataSet, Mul
 
     @Override
     protected int sizeOf(MultiDataSet of) {
-        return of.getFeatures(0).size(0);
+        // FIXME: int cast
+        return  (int) of.getFeatures(0).size(0);
     }
 
     @Override
